Refactor modals project structure
